Tests for DriverManager

+++ src/carrental/DriverManager.java
@@ -37,7 +37,7 @@ public class DriverManager implements IDriverManager{
 	}
 	
 	@Override
-	public Driver findDriverById(int id){
+	public Driver findDriverById(Long id){
 		return new Driver();
 	}
